cell

+++ models/eidetic_3d_lstm_net.py
@@ -47,7 +47,7 @@ def rnn(images, real_input_flag, num_layers, num_hidden, configs):
             kernel=[2, 5, 5])
         lstm_layer.append(new_lstm)
         zero_state = np.zeros(
-            [batch_size, window_length, ims_width, ims_height, num_hidden[i]])
+            [batch_size, window_length, ims_width, ims_height, num_hidden[i]], dtype='float32')
         cell.append(zero_state)
         hidden.append(zero_state)
         c_history.append(None)
